refactor: extract render function and use generate-path of react-router-dom
