+++ src/app/pages/Home/components/SidebarContent/SidebarContent.tsx
@@ -1,11 +1,8 @@
-import { Link, Navigate } from 'react-router-dom';
-import { Box, IconButton, Spinner, Text } from '@chakra-ui/react';
-
 import { GoPlus } from 'react-icons/go';
-import { RxOpenInNewWindow } from 'react-icons/rx';
+import { Box, Spinner, Text } from '@chakra-ui/react';
+import { Navigate, generatePath } from 'react-router-dom';
 
 import { routes } from '@/app/routes';
-import { NoMessagesIcon } from '@/assets';
 
 import { styles } from './styles';
 import { messages } from './messages';
@@ -15,17 +12,19 @@ export const SidebarContent = () => {
   const {
     bots,
     botId,
-    location,
     isPending,
     isHomeRoute,
     isChatbotRoute,
     isChatbotDetailRoute,
     t,
+    renderBots,
     onCreateBot,
   } = usePrepareHook();
 
   if (bots?.length && (isHomeRoute || isChatbotRoute)) {
-    return <Navigate to={routes.chatbotDetail.replace(':id', bots[0].id)} />;
+    return (
+      <Navigate to={generatePath(routes.chatbotDetail, { id: bots[0].id })} />
+    );
   }
 
   if (isHomeRoute) {
@@ -47,47 +46,7 @@ export const SidebarContent = () => {
         <Text fontSize="sm">{t(messages.createChatbot)}</Text>
         {!isPending ? <GoPlus fontSize="1.25rem" /> : <Spinner size="sm" />}
       </Box>
-      <Box marginTop="1rem">
-        {!!bots?.length ? (
-          bots.map(bot => {
-            const isActive = location.pathname.includes(bot.id);
-            return (
-              <Box
-                as={Link}
-                key={bot.id}
-                to={routes.chatbotDetail.replace(':id', bot.id)}
-                background={isActive ? 'gray.100' : 'transparent'}
-                sx={{
-                  ...styles.chatbotContainer,
-                  _hover: {
-                    ...styles.chatbotContainer._hover,
-                    backgroundColor: isActive ? 'gray.100' : 'gray.50',
-                  },
-                }}
-              >
-                <Text fontSize="sm" fontWeight={500}>
-                  {bot.name}
-                </Text>
-
-                <IconButton
-                  id="open-studio-btn"
-                  colorScheme="blue"
-                  aria-label="Open Studio"
-                  sx={styles.openStudioBtn}
-                  icon={<RxOpenInNewWindow fontSize="0.875rem" />}
-                />
-              </Box>
-            );
-          })
-        ) : (
-          <>
-            <NoMessagesIcon width="100%" />
-            <Text sx={styles.emptyText}>
-              {t(messages.thisWorkspaceDoesntHaveChatbotsYet)}
-            </Text>
-          </>
-        )}
-      </Box>
+      <Box marginTop="1rem">{renderBots()}</Box>
     </>
   );
 };
